Feat: 프로필 수정 화면 지역코드 수정 추가

+++ Front/PetPlace/app/src/main/java/com/example/petplace/data/model/mypage/ProfileUpdateRequest.kt
@@ -5,5 +5,5 @@ data class ProfileUpdateRequest(
     val curPassword: String?,
     val newPassword: String?,
     val imgSrc: String?,
-    val regionId: Int?
+    val regionId: Long?
 )
